use RawCatalogItem in cloudrift provider move test to integrity tests add cloudrift to internal catalog
