+++ src/gpuhunt/providers/cloudrift.py
@@ -20,13 +20,13 @@ class CloudRiftProvider(AbstractProvider):
     ) -> list[RawCatalogItem]:
         instance_types = self._get_instance_types()
         instance_types = [inst for instance in instance_types for inst in generate_instances(instance)]
-        return sorted(instance_types, key=lambda x: x['price'])
+        return sorted(instance_types, key=lambda x: x.price)
 
     def _get_instance_types(self):
         return self.rift_client.instance_types.list(services=['vm'])
 
 
-def generate_instances(instance) -> list[dict]:
+def generate_instances(instance) -> list[RawCatalogItem]:
     instance_gpu_brand = instance['brand_short']
     dstack_gpu_name = GPU_MAP.get(instance_gpu_brand)
     if dstack_gpu_name is None:
@@ -36,13 +36,14 @@ def generate_instances(instance) -> list[dict]:
     instance_types = []
     for variant in instance['variants']:
         for (location, _count) in variant['available_nodes_per_dc'].items():
-            raw = dict(
+            raw = RawCatalogItem(
                 instance_name=variant['name'],
                 location=location,
                 spot=False,
                 price=variant['cost_per_hour'] / 100,
                 cpu=variant['cpu_count'],
                 memory=variant['dram'] / 1024**3,
+                disk_size=variant['disk'] / 1024**3,
                 gpu_count=variant['gpu_count'],
                 gpu_name=dstack_gpu_name,
                 gpu_memory=round(variant['vram'] / 1024**3),
